Fix wazuh_conf and add disable_max_size function
This commit also makes some fixes in some tests

+++ tests/integration/test_fim/test_report_changes/test_file_size_default.py
@@ -88,7 +88,7 @@ def test_file_size_default(tags_to_apply, filename, folder, get_configuration, c
         diff_file_path = os.path.join(diff_file_path, folder.strip('/'), filename, 'last-entry.gz')
 
     # Create file with a smaller size than the configured value
-    to_write = generateString(size_limit / 2, '0')
+    to_write = generateString(int(size_limit / 2), '0')
     create_file(REGULAR, folder, filename, content=to_write)
 
     check_time_travel(scheduled)
